Update script to deal with `lost+found` directory.
When mounting certain volumes, OpenDJ installation will fail to install as the mounted volume may have some residual information for some reason (i.e. Amazon ElasticBlockStorage's "lost+found" directory). This only occurs on the first installation and is because the OpenDJ setup requires the `/opt/opendj/config/` directory to be empty. Otherwise the volume can be used as a successfully deployed persistent disk.Here we check if there is an `/opt/opendj/config/schema` directory with files signalling that OpenDJ has already been successfully deployed and will launch as expected. If not, we will make sure the `lost+found` directory, if it exists, is deleted before continuing setup.